+++ mopidy/backends/base/playback.py
@@ -287,11 +287,9 @@ class BasePlaybackController(object):
         Typically called by :class:`mopidy.process.CoreProcess` after a message
         from a library thread is received.
         """
-        next_cp_track = self.cp_track_at_eot
-        if next_cp_track is not None and self._next(next_cp_track[1]):
-            original_cp_track = self.current_cp_track
-            self.current_cp_track = next_cp_track
-            self.state = self.PLAYING
+        original_cp_track = self.current_cp_track
+        if self.cp_track_at_eot:
+            self.play(self.cp_track_at_eot)
 
             if self.consume:
                 self.backend.current_playlist.remove(cpid=original_cp_track[0])
@@ -302,47 +300,42 @@ class BasePlaybackController(object):
             self.stop()
             self.current_cp_track = None
 
-    def new_playlist_loaded_callback(self):
+    def on_current_playlist_change(self):
         """
-        Tell the playback controller that a new playlist has been loaded.
+        Tell the playback controller that the current playlist has changed.
 
-        Typically called by :class:`mopidy.process.CoreProcess` after a message
-        from a library thread is received.
+        Used by :class:`mopidy.backends.base.BaseCurrentPlaylistController`.
         """
-        self.current_cp_track = None
         self._first_shuffle = True
         self._shuffled = []
 
-        if self.state == self.PLAYING:
-            if len(self.backend.current_playlist.tracks) > 0:
-                self.play()
-            else:
-                self.stop()
-        elif self.state == self.PAUSED:
+        if not self.backend.current_playlist.cp_tracks:
+            self.stop()
+            self.current_cp_track = None
+        elif (self.current_cp_track not in
+                self.backend.current_playlist.cp_tracks):
+            self.current_cp_track = None
             self.stop()
 
     def next(self):
         """Play the next track."""
-        original_cp_track = self.current_cp_track
-
         if self.state == self.STOPPED:
             return
-        elif self.cp_track_at_next is not None and self._next(self.next_track):
-            self.current_cp_track = self.cp_track_at_next
-            self.state = self.PLAYING
-        elif self.cp_track_at_next is None:
+
+        original_cp_track = self.current_cp_track
+        if self.cp_track_at_next:
+            self.play(self.cp_track_at_next)
+        else:
             self.stop()
             self.current_cp_track = None
 
-        # FIXME handle in play aswell?
+        # FIXME This should only be applied when reaching end of track, and not
+        # when pressing "next"
         if self.consume:
             self.backend.current_playlist.remove(cpid=original_cp_track[0])
 
         if self.random and self.current_cp_track in self._shuffled:
             self._shuffled.remove(self.current_cp_track)
-
-    def _next(self, track):
-        return self._play(track)
 
     def pause(self):
         """Pause playback."""
@@ -352,13 +345,16 @@ class BasePlaybackController(object):
     def _pause(self):
         raise NotImplementedError
 
-    def play(self, cp_track=None):
+    def play(self, cp_track=None, on_error_step=1):
         """
         Play the given track or the currently active track.
 
         :param cp_track: track to play
         :type cp_track: two-tuple (CPID integer, :class:`mopidy.models.Track`)
             or :class:`None`
+        :param on_error_step: direction to step at play error, 1 for next
+            track (default), -1 for previous track
+        :type on_error_step: int, -1 or 1
         """
 
         if cp_track is not None:
@@ -368,13 +364,14 @@ class BasePlaybackController(object):
 
         if self.state == self.PAUSED and cp_track is None:
             self.resume()
-        elif cp_track is not None and self._play(cp_track[1]):
+        elif cp_track is not None:
             self.current_cp_track = cp_track
             self.state = self.PLAYING
-
-        # TODO Do something sensible when _play() returns False, like calling
-        # next(). Adding this todo instead of just implementing it as I want a
-        # test case first.
+            if not self._play(cp_track[1]):
+                if on_error_step == 1:
+                    self.next()
+                elif on_error_step == -1:
+                    self.previous()
 
         if self.random and self.current_cp_track in self._shuffled:
             self._shuffled.remove(self.current_cp_track)
@@ -384,14 +381,11 @@ class BasePlaybackController(object):
 
     def previous(self):
         """Play the previous track."""
-        if (self.previous_cp_track is not None
-                and self.state != self.STOPPED
-                and self._previous(self.previous_track)):
-            self.current_cp_track = self.previous_cp_track
-            self.state = self.PLAYING
-
-    def _previous(self, track):
-        return self._play(track)
+        if self.previous_cp_track is None:
+            return
+        if self.state == self.STOPPED:
+            return
+        self.play(self.previous_cp_track, on_error_step=-1)
 
     def resume(self):
         """If paused, resume playing the current track."""